GroupBy: Check if the file is created before trying to delete it.
git-svn-id: svn://source.pentaho.org/svnkettleroot/mirror/Kettle/trunk@60 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/be/ibridge/kettle/trans/step/groupby/GroupBy.java
@@ -498,7 +498,7 @@ public class GroupBy extends BaseStep implements StepInterface
 		}
 		finally
 		{
-            data.tempFile.delete();
+            if (data.tempFile!=null) data.tempFile.delete();
 			dispose(meta, data);
 			logSummary();
 			markStop();
